CuratorAgent/requirements.txt
qdrant-client>=1.7.0
numpy>=1.24.0
hdbscan>=0.8.33
scikit-learn>=1.3.0
openai>=1.12.0
jinja2>=3.1.0
python-dotenv>=1.0.0
agency-swarm>=0.1.0 